@@ -23,7 +23,7 @@
     }
   },
   "scripts": {
-    "test": "jest --coverage",
+    "test": " node --experimental-vm-modules node_modules/.bin/jest --coverage",
     "test:watch": "jest --watch --verbose --coverage",
     "clean": "rimraf coverage/ lib/",
     "build": "rollup -c rollup.config.js && cp ../../README.md ./",
@@ -33,43 +33,32 @@
   },
   "license": "BSD-3-Clause",
   "dependencies": {
-    "cookie": "^0.3.1",
+    "cookie": "^0.5.0",
     "escape-html": "^1.0.3",
     "intl": "^1.2.5",
-    "intl-messageformat": "^7.8.4",
-    "invariant": "^2.2.2",
-    "lodash.merge": "^4.6.2",
-    "object-keys": "^1.0.11",
-    "querystring": "^0.2.0"
+    "intl-messageformat": "^10.3.1",
+    "invariant": "^2.2.4",
+    "plain-object-merge": "^2.0.0"
   },
+  "type": "module",
   "peerDependencies": {
     "react": "*"
   },
   "devDependencies": {
-    "babel-cli": "^6.2.0",
-    "babel-eslint": "^7.1.1",
-    "babel-jest": "^19.0.0",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
-    "babel-preset-es2015": "^6.1.18",
-    "babel-preset-es2015-rollup": "^3.0.0",
-    "jest": "^19.0.2",
-    "jest-cli": "^19.0.2",
-    "react": "^15.4.2",
+    "@babel/core": "^7.21.0",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@rollup/plugin-babel": "^6.0.3",
+    "jest": "^29.5.0",
+    "jest-cli": "^29.5.0",
+    "babel-jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "react": "^18.2.0",
     "rimraf": "^2.4.2",
-    "rollup": "^0.41.4",
-    "rollup-plugin-babel": "^2.3.9",
-    "rollup-plugin-commonjs": "^7.0.0"
-  },
-  "babel": {
-    "presets": [
-      "es2015",
-      "react"
-    ]
+    "rollup": "^3.19.1"
   },
   "jest": {
-    "transform": {
-      ".*": "<rootDir>/node_modules/babel-jest"
-    },
+    "testEnvironment": "jsdom",
     "testRegex": "/test/.*\\.test.js",
     "testPathIgnorePatterns": [],
     "collectCoverageFrom": [
